refactor: [EL-4696] extract notification helpers and message components into FSD entities layer
