+++ src/[fsd]/entities/notifications/lib/helpers/index.js
@@ -1,0 +1,3 @@
+export * from './notification.helpers.js';
+// Legacy: remove once all environments have run the backfill migration
+export * from './notification.legacy-helpers.js';
