+++ src/[fsd]/entities/notifications/ui/index.js
@@ -1,0 +1,2 @@
+export { default as LegacyNotificationMessage } from './LegacyNotificationMessage.jsx';
+export { default as NotificationListItemMessage } from './NotificationListItemMessage.jsx';
